Polishing etcd backend.
Applying some fixes and going simple now on the
EtcdCoordinator. Added some tests to verify that
commands are delivered in the same order even when
multiple replicas exists.

After some digging, find out that there is no need
to use the KV to send requests, this can be done with
the ETCD client directly. So this was removed, and also
there is no need to synchronously send commands nor treat
watch events asynchonously.

Added a load benchmark test to be verified with `pprof`. Any
optimization will be done later.